Extract laps_search from smb and winrm to /protocols/ldap/laps
